requirements.txt
langchain==0.3.27
langchain_community==0.3.29
langchain_openai==0.3.33
pandas==2.3.2
pydantic==2.11.9
python-dotenv==1.1.1
streamlit==1.49.1
